+++ tensor_io/process_frostt_tensor.cpp
@@ -15,7 +15,7 @@ using namespace std;
  * This file assumes that the tensor is 1-indexed. 
  */
 
-void convertFromFROSTT(string in_file, unsigned long long num_lines) {
+void convertFromFROSTT(string in_file, uint64_t num_lines) {
     cout << "Starting file conversion!" << endl; 
 
     string converted_filename = in_file + "_converted.hdf5";
@@ -57,9 +57,9 @@ void convertFromFROSTT(string in_file, unsigned long long num_lines) {
     hid_t min_mode_dataset = H5Dcreate(file, min_mode_set.c_str(), idx_datatype, mode_size_dataspace,
                 H5P_DEFAULT, H5P_DEFAULT, H5P_DEFAULT); 
 
-    vector<unique_ptr<unsigned long long>> idx_buffers;
-    vector<unsigned long long> mode_maxes;
-    vector<unsigned long long> mode_mins;
+    vector<unique_ptr<uint64_t>> idx_buffers;
+    vector<uint64_t> mode_maxes;
+    vector<uint64_t> mode_mins;
 
     vector<hid_t> datasets;
 
@@ -70,7 +70,7 @@ void convertFromFROSTT(string in_file, unsigned long long num_lines) {
 
 
     for(int i = 0; i < dim; i++) {
-        idx_buffers.emplace_back(new unsigned long long[BUFFER_SIZE]);
+        idx_buffers.emplace_back(new uint64_t[BUFFER_SIZE]);
 
         string datasetname = "MODE_" + std::to_string(i);
 
@@ -91,11 +91,11 @@ void convertFromFROSTT(string in_file, unsigned long long num_lines) {
     hsize_t buffer_size = BUFFER_SIZE;
     hid_t memory_dataspace = H5Screate_simple(1, &buffer_size, NULL); 
 
-    unsigned long long pos_in_buffer = 0;
+    uint64_t pos_in_buffer = 0;
 
     for(hsize_t i = 0; i < num_lines; i++) {
         for(int j = 0; j < dim; j++) {
-            unsigned long long idx;
+            uint64_t idx;
             iffstream >> idx;
             idx_buffers[j].get()[pos_in_buffer] = idx;
 
@@ -179,7 +179,7 @@ void convertFromFROSTT(string in_file, unsigned long long num_lines) {
 
 int main(int* argc, char** argv) {
     string name(argv[1]);
-    unsigned long long num_lines = atol(argv[2]);
+    uint64_t num_lines = atol(argv[2]);
  
     convertFromFROSTT(name, num_lines);
 }
